Bug 🐛: Fix infinite scroll
Fixed the issue with infinite scroll not working. It was do with my
custom react hook not working properly. Installed library that checks
when the scroll intersection and then loads 10 new tweets based on that.

Fixed a whole boat load of bugs:
    - Removed extraneous file imports that had become unnecessary.
    - Used isomorphic dompurify over regular dom purify for nextjs.
    - Added tweet feed component that can be used on other pages as well

Big Bug still to fix: HOW TO RENDER USER CONTENT FROM A RICH TEXT EDITOR
!!!!!!!!!!!!!!!!!!!!!!!!!!!

IT'S GOD DAMN HARD.